pyproject.toml
[tool.poetry]
name = "leetcode-scraper"
version = "1.0.0"
description = "Fetch and store LeetCode study plan details."
authors = ["Nicholas Adamou <nicholasadamouemail@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pyfiglet = "^1.0.2"
toml = "^0.10.2"
pyyaml = "^6.0.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pytest = "^8.2.2"
pre-commit = "^3.7.1"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
